Moves association proxy tests to their own module.
